+++ src/info/kanlaki101/blockprotection/commands/BPAdmin.java
@@ -1,6 +1,7 @@
 package info.kanlaki101.blockprotection.commands;
 
 import info.kanlaki101.blockprotection.BlockProtection;
+import info.kanlaki101.blockprotection.utilities.BPConfigHandler;
 
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
@@ -18,7 +19,7 @@ public class BPAdmin implements CommandExecutor {
 		String noperm = "You do not have permission to use this command.";
 		ChatColor YELLOW = ChatColor.YELLOW;
 		
-		if (commandLabel.equalsIgnoreCase("bpadmin")) {
+		if (cmd.getName().equalsIgnoreCase("bpadmin")) {
 			if (!(sender instanceof Player)) return true;
 			
 			Player p = (Player) sender;
@@ -26,18 +27,18 @@ public class BPAdmin implements CommandExecutor {
 			String bpp = pl.prefix + player;
 			if (!pl.isAuthorized(p, "bp.admin")) { //No permissions
 				p.sendMessage(YELLOW + noperm);
-				if (pl.advLog()) pl.log.warning(bpp + " attempted to use command: 'bpadmin'."); //Log attempted use of command
+				if (BPConfigHandler.advLog() == true) pl.log.warning(bpp + " attempted to use command: 'bpadmin'."); //Log attempted use of command
 				return true;
 			}
 			if (pl.UsersBypass.contains(player)) { //Check for the player in the bypass protection array
 				pl.UsersBypass.remove(player); //Remove them from the array. (Can't break blocks)
 				p.sendMessage(YELLOW + "You can no longer break other players blocks.");
-				if (pl.advLog()) pl.log.info(bpp + " is no longer bypassing BlockProtection"); //Log to console
+				if (BPConfigHandler.advLog() == true) pl.log.info(bpp + " is no longer bypassing BlockProtection"); //Log to console
 			}
 			else {
 				pl.UsersBypass.add(player); //Add them to the array, and allow them to break blocks
 				p.sendMessage(YELLOW + "You can now break other players blocks.");
-				if (pl.advLog()) pl.log.info(bpp + " is now bypassing BlockProtection"); //Log to console
+				if (BPConfigHandler.advLog() == true) pl.log.info(bpp + " is now bypassing BlockProtection"); //Log to console
 			}
 		}
 		return true;