+++ src/info/kanlaki101/blockprotection/commands/BPClear.java
@@ -1,6 +1,7 @@
 package info.kanlaki101.blockprotection.commands;
 
 import info.kanlaki101.blockprotection.BlockProtection;
+import info.kanlaki101.blockprotection.utilities.BPConfigHandler;
 
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
@@ -21,21 +22,22 @@ public class BPClear implements CommandExecutor {
 		String noperm = "You do not have permission to use this command.";
 		ChatColor YELLOW = ChatColor.YELLOW;
 		
-		if (args.length > 1) return true; //Too many arugments
-		if (commandLabel.equalsIgnoreCase("bpclear")) {
-			pl.loadFriendsList();
+		if (args.length > 1) return true; //Too many arguments
+		
+		if (cmd.getName().equalsIgnoreCase("bpclear")) {
+			BPConfigHandler.loadFriendsList();
 			if (args.length > 0) return true;
 			if (!pl.isAuthorized(p, "bp.friend")) { //No permissions
 				p.sendMessage(YELLOW + noperm);
 				return true;
 			}
-			if (pl.friendslist.getList(player) == null) { //Check if player has a friends list
+			if (BPConfigHandler.getFriendslist(player) == null) { //Check if player has a friends list
 				p.sendMessage(YELLOW + "You do not have a friends list.");
 			}
 			else {
-				pl.friendslist.set(player, null); //Remove everyone from the list
+				BPConfigHandler.friendslist.set(player, null); //Remove everyone from the list
 				p.sendMessage(YELLOW + "Your friends list has been cleared.");
-				pl.saveFriendsList(); //Save friends list
+				BPConfigHandler.saveFriendsList(); //Save friends list
 			}
 		}
 	return true;