+++ src/info/kanlaki101/blockprotection/commands/BPRemove.java
@@ -1,6 +1,7 @@
 package info.kanlaki101.blockprotection.commands;
 
 import info.kanlaki101.blockprotection.BlockProtection;
+import info.kanlaki101.blockprotection.utilities.BPConfigHandler;
 
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
@@ -21,15 +22,15 @@ public class BPRemove implements CommandExecutor {
 		String noperm = "You do not have permission to use this command.";
 		ChatColor YELLOW = ChatColor.YELLOW;
 		
-		if (args.length > 1) return true; //Too many arugments
+		if (args.length > 1) return true; //Too many arguments
 
-		if (commandLabel.equalsIgnoreCase("bpremove")) {
-			pl.loadFriendsList();
+		if (cmd.getName().equalsIgnoreCase("bpremove")) {
+			BPConfigHandler.loadFriendsList();
 			if (!pl.isAuthorized(p, "bp.friend")) { //No permissions
 				p.sendMessage(YELLOW + noperm);
 				return true;
 			}
-			if (pl.friendslist.getList(player) == null) { //Check if a friends list exist for them
+			if (BPConfigHandler.friendslist.getList(player) == null) { //Check if a friends list exist for them
 				p.sendMessage(YELLOW + "You do not have a friends list.");
 				return true;
 			}
@@ -38,12 +39,12 @@ public class BPRemove implements CommandExecutor {
 				return true;
 			}
 			
-			if (pl.friendslist.getList(player).contains(args[0])) { //If player is already in your friends list
-				pl.friendslist.getList(player).remove(args[0]); //Remove him
-		        if (pl.friendslist.getList(player).isEmpty()) { //If the list is empty
-		        	pl.friendslist.set(player, null); //Delete it
+			if (BPConfigHandler.getFriendslist(player).contains(args[0])) { //If player is already in your friends list
+				BPConfigHandler.getFriendslist(player).remove(args[0]); //Remove him
+		        if (BPConfigHandler.getFriendslist(player).isEmpty()) { //If the list is empty
+		        	BPConfigHandler.friendslist.set(player, null); //Delete it
 		        }
-		        pl.saveFriendsList(); //Save
+		        BPConfigHandler.saveFriendsList(); //Save
 		        p.sendMessage(YELLOW + args[0] + " has been removed from your friends list.");
 			} 
 			else {