+++ src/info/kanlaki101/blockprotection/commands/BPTool.java
@@ -1,8 +1,10 @@
 package info.kanlaki101.blockprotection.commands;
 
 import info.kanlaki101.blockprotection.BlockProtection;
+import info.kanlaki101.blockprotection.utilities.BPConfigHandler;
 
 import org.bukkit.ChatColor;
+import org.bukkit.Material;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -19,20 +21,37 @@ public class BPTool implements CommandExecutor {
 		String noperm = "You do not have permission to use this command.";
 		ChatColor YELLOW = ChatColor.YELLOW;
 		
-		if (commandLabel.equalsIgnoreCase("bptool")) {
+		if (cmd.getName().equalsIgnoreCase("bptool")) {
 			if (!(sender instanceof Player)) return true;
 			
 			Player p = (Player) sender;
 			String player = p.getName();
 			String bpp = pl.prefix + player;
+			
+			BPConfigHandler.loadConfig();
 			if (!pl.isAuthorized(p, "bp.admin")) { //No permissions
 				p.sendMessage(YELLOW + noperm);
-				if (pl.advLog()) pl.log.warning(bpp + " attempted to use command: 'bptool'.");
+				if (BPConfigHandler.advLog() == true) pl.log.warning(bpp + " attempted to use command: 'bptool'.");
 			} 
 			else {
-				p.getInventory().addItem(new ItemStack(pl.config.getInt("utilitytool"), 1 )); //Give player a "utility tool"
-				p.sendMessage(YELLOW + "Utility Tool added to inventory.");
-				if (pl.advLog()) pl.log.info(bpp + " has been given a utility tool."); //Log to console
+				if (args.length > 1) return true; //Too many arguments
+				
+				if (p.getInventory().firstEmpty() == -1) { //If the player's inventory is full
+					p.sendMessage(YELLOW + "Can not add tool. Inventory is full.");
+					return true;
+				}
+				
+				Material item = Material.getMaterial(BPConfigHandler.getUtilTool());
+				if (p.getInventory().contains(item)) {
+					p.sendMessage(YELLOW + "You already have a " + item + ".");
+					return true;
+				}
+				
+				p.getInventory().addItem(new ItemStack(item, 1 ));
+				p.sendMessage(YELLOW + "" + item + " has been added to your inventory.");
+				
+				if (BPConfigHandler.advLog() == true) pl.log.info(bpp + " has been given a utility tool."); //Log to console
+				return true;
 			}
 		}
 		return true;